backend/requirements.txt
flask
python-dotenv
spacy
python-docx
pypdf2
flask-cors


